+++ packages/client3/.eslintrc.json
@@ -5,7 +5,7 @@
   "rules": {
     "no-console": ["warn", { "allow": ["warn", "error"] }],
     "react-hooks/exhaustive-deps": "error",
-    "@typescript-eslint/no-unused-vars": "error",
+    "@typescript-eslint/no-unused-vars": "warn",
     "import/order": [
       "error",
       {
